+++ packages/ra-data-simple-prisma/src/extractWhere.ts
@@ -1,4 +1,10 @@
-import { PlainObject, isObject, setObjectPath } from "deverything";
+import {
+  PlainObject,
+  isArray,
+  isObject,
+  isString,
+  setObjectPath,
+} from "deverything";
 import { GetListRequest, GetManyReferenceRequest } from "./Http";
 import { isNotField } from "./lib/isNotField";
 
@@ -19,6 +25,7 @@ export type FilterMode = "insensitive" | "default" | undefined;
 
 type ExtractWhereOptions = {
   filterMode?: FilterMode;
+  debug?: boolean;
 };
 
 export const extractWhere = (
@@ -27,21 +34,27 @@ export const extractWhere = (
 ) => {
   const { filter } = req.params;
 
+  console.debug("extractWhere:filter", filter);
+
   const where = {};
 
-  if (filter) {
-    Object.entries(filter).forEach(([colName, value]) => {
-      if (isNotField(colName)) return;
+  const setWhere = (filter: PlainObject, currentFilterPath?: string) => {
+    Object.entries(filter).forEach(([field, value]) => {
+      if (isNotField(field)) return;
 
       //TODO: *consider* to move into `isNotField` (but maybe to reset dates is the only way to do it)
       if (value === "")
         //react-admin does send empty strings in empty filters :(
         return;
 
+      const filterPath = currentFilterPath
+        ? `${currentFilterPath}.${field}`
+        : field;
+
       const hasOperator = prismaOperators.some((operator) => {
-        if (colName.endsWith(`_${operator}`)) {
-          [colName] = colName.split(`_${operator}`);
-          setObjectPath(where, colName, { [operator]: value });
+        if (field.endsWith(`_${operator}`)) {
+          const [wherePath] = filterPath.split(`_${operator}`);
+          setObjectPath(where, wherePath + `.${operator}`, value);
           return true;
         }
       });
@@ -49,45 +62,55 @@ export const extractWhere = (
 
       if (
         // Custom operators
-        colName.endsWith(`_enum`) ||
-        colName.endsWith(`_exact`) ||
-        colName.endsWith(`_eq`)
+        field.endsWith(`_enum`) ||
+        field.endsWith(`_exact`) ||
+        field.endsWith(`_eq`)
       ) {
-        const [cleanColName] = colName.split(/(_enum|_exact|_eq)$/);
-        setObjectPath(where, cleanColName, value);
-      } else if (colName === "q") {
-        // i.e. when filterToQuery is not set on AutoCompleteInput, but we don't know all the fields to search against
-        console.info("Filter not handled:", colName, value);
-      } else if (
-        colName === "id" ||
-        colName === "uuid" ||
-        colName === "cuid" ||
-        colName.endsWith("_id") ||
-        colName.endsWith("Id") ||
-        typeof value === "number" ||
-        typeof value === "boolean" ||
-        value === null // if the client sends null, than that is also a valid (exact) filter!
-      ) {
-        setObjectPath(where, colName, value);
-      } else if (Array.isArray(value)) {
-        setObjectPath(where, colName, { in: value });
-      } else if (typeof value === "string") {
-        setObjectPath(where, colName, {
-          contains: value,
-          mode: options?.filterMode,
-        });
-      } else if (isObject(value)) {
+        const [wherePath] = filterPath.split(/(_enum|_exact|_eq)$/);
+        setObjectPath(where, wherePath, value);
+      } else if (field.endsWith(`_pgjson`)) {
+        const [wherePath] = filterPath.split("_pgjson");
+
         // if object then it's a Json field, this is EXPERIMENTAL and works only for Postgres
         // https://www.prisma.io/docs/concepts/components/prisma-client/working-with-fields/working-with-json-fields#filter-on-object-property
         const { path, equals } = formatPrismaPostgresNestedJsonFilter(value);
         if (path.length && equals) {
-          setObjectPath(where, colName, { path, equals });
+          setObjectPath(where, wherePath, { path, equals }); // TODO: allow operators
         }
+      } else if (field === "q") {
+        // i.e. when filterToQuery is not set on AutoCompleteInput, but we don't know all the fields to search against
+        console.info("Filter not handled:", field, value);
+      } else if (
+        field === "id" || // careful not to use filterPath here
+        field === "uuid" ||
+        field === "cuid" ||
+        field.endsWith("_id") ||
+        field.endsWith("Id") ||
+        typeof value === "number" ||
+        typeof value === "boolean" ||
+        value === null // if the client sends null, than that is also a valid (exact) filter!
+      ) {
+        setObjectPath(where, filterPath, value);
+      } else if (isArray(value)) {
+        setObjectPath(where, filterPath, { in: value });
+      } else if (isString(value)) {
+        setObjectPath(where, filterPath, {
+          contains: value,
+          mode: options?.filterMode,
+        });
+      } else if (isObject(value)) {
+        setWhere(value, filterPath); // Recursively handle nested objects
       } else {
-        console.info("Filter not handled:", colName, value);
+        console.info("ra-data-simple-prisma: Filter not handled", field, value);
       }
     });
+  };
+
+  if (filter) {
+    setWhere(filter);
   }
+
+  console.debug("extractWhere:where", where);
 
   return where;
 };